requirements.txt
requests 
beautifulsoup4 
psycopg2
python-dotenv
google-generativeai
google-genai


